Move header to its own file
