+++ src/services/AbstractService.spec.ts
@@ -26,7 +26,9 @@ const setMetadata = () => (_metadata: Metadata) => 'setMetadata set!';
 const register = (..._args: unknown[]) => 'registered!';
 
 /**
- * Mock polkadot-js ApiPromise.
+ * Mock polkadot-js ApiPromise. This has its own mock api for now just for
+ * simplicity sake. In the future this may be replaced by
+ * the mockApi in ./utils/mock
  */
 const api = {
 	createType: kusamaRegistry.createType.bind(kusamaRegistry),
