+++ src/services/utils/header78629.ts
@@ -1,0 +1,29 @@
+/**
+ * Header from polkadot block #789629
+ */
+export const header789629 = {
+	parentHash:
+		'0x205da5dba43bbecae52b44912249480aa9f751630872b6b6ba1a9d2aeabf0177',
+	number: 789629,
+	stateRoot:
+		'0x023b5bb1bc10a1a91a9ef683f46a8bb09666c50476d5592bd6575a73777eb173',
+	extrinsicsRoot:
+		'0x4c1d65bf6b57086f00d5df40aa0686ffbc581ef60878645613b1fc3303de5030',
+	digest: {
+		logs: [
+			{
+				PreRuntime: [
+					1161969986,
+					'0x036d000000f4f4d80f00000000ec9bd8e2d0368c97f3d888837f7283bbe08266869eb613159db547905026c2502a70f168b9ffcc233344005d11ebecd166769200d270a2eaa642118a00acb708a0487a440b0caf3dd5c91ab173e80ddfe5735ef8b938ea87a6105a1161612707',
+				],
+			},
+			{
+				Seal: [
+					1161969986,
+					'0xae78514e1de84a7d32e55b9b652f9d408ab1f7b4bfdbf6b2fad9cad94a91b86b0161cabf08f5ae1d3a1aa4993e2d96d56c94b03cee0898ccb8385a546084f88b',
+				],
+			},
+		],
+	},
+	author: '1zugcajGg5yDD9TEqKKzGx7iKuGWZMkRbYcyaFnaUaEkwMK',
+};
